# docs/openapi/c2c-market-api-v1.yaml
openapi: 3.1.0
info:
  title: C2CMarket API
  version: 0.1.0
  summary: Backend contract foundation and official price lead slice.
  description: |
    This contract covers the first executable backend slice. C2CMarket is a price intelligence
    and community matching platform. It does not process payments, escrow funds, store third-party
    credentials, store plaintext native passwords, auto-deliver keys/accounts, or proxy upstream API
    traffic. Native C2CMarket account passwords are stored only as salted hashes.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
paths:
...
  /api/v1/auth/dev-session:
    post:
      x-dev-only: true
      operationId: createDevSession
      tags: [DevAuth]
      security: []
      summary: Create a local development session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                username:
                  type: string
                  maxLength: 64
                admin:
                  type: boolean
      responses:
        "200":
          description: Development session.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/contact-methods:
    post:
      operationId: createContactMethod
      tags: [ContactMethods]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactMethodRequest"
      responses:
        "201":
          description: Contact method created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMethod"
  /api/v1/contact-methods/{id}:
    patch:
      operationId: updateContactMethod
      tags: [ContactMethods]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactMethodRequest"
      responses:
        "200":
          description: Contact method updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMethod"
    delete:
      operationId: deleteContactMethod
      tags: [ContactMethods]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Contact method disabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMethod"
  /api/v1/contact-methods/{id}/set-default:
    post:
      operationId: setDefaultContactMethod
      tags: [ContactMethods]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Contact method set as default.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMethod"
  /api/v1/contact-methods/{id}/verify:
    post:
      operationId: verifyContactMethod
      tags: [ContactMethods]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Contact method marked verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMethod"
  /api/v1/dev/contact-sessions:
    post:
      x-dev-only: true
      operationId: createDevContactSession
      tags: [DevContact]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDevContactSessionRequest"
      responses:
        "201":
          description: Contact session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSession"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: c2c_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    RateLimited:
      description: Rate limit exceeded. Problem Details `code` is `RATE_LIMITED`.
      headers:
        Retry-After:
          schema:
            type: string
          description: Seconds until the current in-process rate-limit window resets when known.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    SessionResponse:
      type: object
      required: [user, csrfToken, expiresAt]
      properties:
        user:
          $ref: "#/components/schemas/User"
        csrfToken:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    LinuxDoBinding:
      type: object
      required: [bound]
      properties:
        bound:
          type: boolean
        linuxDoUserId:
          type: string
        linuxDoUsername:
          type: string
        trustLevel:
          type: integer
        avatarUrl:
          type: string
          format: uri
    User:
      type: object
      required: [id, username, displayName, isAdmin, permissions, linuxDoBinding]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        isAdmin:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        linuxDoBinding:
          $ref: "#/components/schemas/LinuxDoBinding"
...
    CreateContactMethodRequest:
      type: object
      additionalProperties: false
      required: [type, label]
      properties:
        type:
          type: string
          enum: [linuxdo, telegram, wechat, email, other]
        label:
          type: string
        value:
          type: string
        displayValue:
          type: string
        usageScopes:
          type: array
          items:
            type: string
        isDefault:
          type: boolean
        enabled:
          type: boolean
    ContactMethod:
      type: object
      required: [id, userId, type, label, maskedValue, usageScopes, isDefault, enabled, verified, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
        label:
          type: string
        maskedValue:
          type: string
        displayValue:
          type: string
          description: Returned only in authorized self/contact-window contexts when implemented.
        usageScopes:
          type: array
          items:
            type: string
        isDefault:
          type: boolean
        enabled:
          type: boolean
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
...
    CreateDevContactSessionRequest:
      type: object
      additionalProperties: false
      required: [sellerUsername, buyerContactMethodId, sellerContactMethodId]
      properties:
        sellerUsername:
          type: string
        buyerContactMethodId:
          type: string
          format: uuid
        sellerContactMethodId:
          type: string
          format: uuid
        durationSeconds:
          type: integer
          minimum: 1
    ContactSession:
      type: object
      required: [id, buyerId, sellerId, opensAt, endsAt]
      properties:
        id:
          type: string
          format: uuid
        buyerId:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        opensAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, detail, instance, requestId]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string